商品削除機能の修正(1)

+++ app/controllers/items_controller.rb
@@ -42,6 +42,8 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
+    return unless item.user.id == current_user.id
+
     item.destroy
     redirect_to root_path
   end
